Change attachment size from int to long. Thanks bback.
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/Freetalk@26565 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/plugins/Freetalk/Message.java
@@ -101,9 +101,9 @@ public abstract class Message implements Comparable<Message> {
 	
 	public static class Attachment {
 		private final FreenetURI mURI;
-		private final int mSize; /* Size in bytes */
+		private final long mSize; /* Size in bytes */
 		
-		public Attachment(FreenetURI myURI, int mySize) {
+		public Attachment(FreenetURI myURI, long mySize) {
 			mURI = myURI;
 			mSize = mySize;
 		}
@@ -112,7 +112,7 @@ public abstract class Message implements Comparable<Message> {
 			return mURI;
 		}
 		
-		public int getSize() {
+		public long getSize() {
 			return mSize;
 		}
 	}
